move From<&QueryTracker> for TaskTracker
